Replace trackRequests Vuex module with FcRequestsTrack-local state
We only use this in one place and can pass data down via props, so there's no reason to have the additional complexity of Vuex here.
